KERRY IT IS IRONY. YOU NEVER EVER COMMIT REGULARY AND SUDDENLY THE MOMENT YOU NEED ME TO I DO IT IMMEDAITELY. WHY? BECUASE LIKE EVERYONE ELSE IN THIS GROUP I DO WHAT IS HELPFUL TO THE TEAM. I AM NOT A SELFISH CUNT WHO THINKS THEY CAN DO WHATEVER THEY WANT. NOT ONLY DOES THIS MAKE YOU A BAD TEAM PLAYER BUT ALSO A BAD PERSON AND BAD PROGRAMMER.
Also, made a start on TestSimoriPanel, its quite long, so it might take
a while.

blumpkin:

The delicately balanced art of getting your cock sucked while taking a
dump. 

Morty made frequent trips to the office restroom, but never when the
janitor who calls Costa Rica is there, because that janitor like to give
Blumpkins.